- Update site characteristics exit activity - Implement autopopulate site settings for edit
