+++ opensrp-anc/src/main/java/org/smartregister/anc/activity/SiteCharacteristicsExitActivity.java
@@ -1,14 +1,14 @@
 package org.smartregister.anc.activity;
 
 import android.content.Intent;
-import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
-import android.support.v7.app.ActionBar;
 import android.util.Log;
 import android.view.View;
+import android.widget.TextView;
 
 import org.smartregister.anc.R;
 import org.smartregister.anc.contract.SiteCharacteristicsContract;
+import org.smartregister.anc.helper.LocationHelper;
 import org.smartregister.anc.presenter.SiteCharacteristicsPresenter;
 import org.smartregister.anc.util.Constants;
 import org.smartregister.anc.util.JsonFormUtils;
@@ -20,7 +20,6 @@ import java.util.Map;
  */
 public class SiteCharacteristicsExitActivity extends BaseActivity implements View.OnClickListener, SiteCharacteristicsContract.View {
     private static final String TAG = SiteCharacteristicsExitActivity.class.getCanonicalName();
-    private SiteCharacteristicsContract.Presenter presenter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -28,14 +27,29 @@ public class SiteCharacteristicsExitActivity extends BaseActivity implements Vie
         setContentView(R.layout.activity_site_characteristics_exit);
 
         findViewById(R.id.btn_site_characteristics_home_register).setOnClickListener(this);
+        findViewById(R.id.btn_back_to_home).setOnClickListener(this);
 
         presenter = new SiteCharacteristicsPresenter(this);
+
+        String defaultLocation = LocationHelper.getInstance().getOpenMrsLocationName(LocationHelper.getInstance().getDefaultLocation());
+        TextView textView = findViewById(R.id.site_characteristics_facility_name);
+        textView.setText(getString(R.string.your_site_characteristics_for_facility_name, defaultLocation));
+
 
     }
 
     @Override
     public void onClick(View view) {
-        goToHomeRegisterPage();
+
+        if (view.getId() == R.id.btn_site_characteristics_home_register) {
+
+            goToHomeRegisterPage();
+
+        } else {
+
+            presenter.launchSiteCharacteristicsFormForEdit();
+        }
+
     }
 
     public void goToHomeRegisterPage() {
@@ -44,12 +58,6 @@ public class SiteCharacteristicsExitActivity extends BaseActivity implements Vie
         startActivity(intent);
 
         finish();//finish this
-    }
-
-    @Override
-    public void onBackPressed() {
-        super.onBackPressed();
-        presenter.launchSiteCharacteristicsFormForEdit();
     }
 
     @Override
