--- opensrp-anc/src/main/res/layout/activity_site_chars_exit_toolbar_bar_content.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/primary"
-    android:gravity="center_vertical"
-    android:orientation="horizontal">
-
-    <ImageView
-        android:id="@+id/btn_back_to_home"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:scaleType="fitXY"
-        android:src="@drawable/ic_back" />
-
-    <org.smartregister.view.customcontrols.CustomFontTextView
-        android:id="@+id/txt_title_label"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:background="@android:color/transparent"
-        android:ellipsize="end"
-        android:gravity="center_vertical"
-        android:singleLine="true"
-        android:text="@string/go_back_to_site_charactersitics"
-        android:textColor="@color/white"
-        android:layout_marginLeft="16dp"/>
-
-
-</LinearLayout>
