+++ opensrp-anc/src/main/res/values/strings.xml
@@ -227,7 +227,7 @@
     <string name="continue_to_your_client_list">CONTINUE TO YOUR CLIENT LIST</string>
     <string name="your_site_characteristics_for_facility_name">Your site characteristics for %1$s are all set.</string>
     <string name="site_characteristics_are_shared_for_all">Site charactersitics ares shared by all OpenSRP users at the facility. You can update site characteristics at any time</string>
-    <string name="go_back_to_site_charactersitics">Go back to site characteristics</string>
+    <string name="go_back_to_site_characteristics">Go back to site characteristics</string>
 
     <!-- contact -->
     <string name="profile">Profile</string>
